Basic collision bounce system

+++ src/ECS/Components/Sprite.cpp
@@ -27,8 +27,8 @@ void Sprite::draw() {
 }
 
 void Sprite::update() {
-    destRect.x = (int) transform->position.x;
-    destRect.y = (int) transform->position.y;
-    destRect.w = transform->width * transform->scale;
-    destRect.h = transform->height * transform->scale;
+    destRect.x = static_cast<int>(transform->position.x);
+    destRect.y = static_cast<int>(transform->position.y);
+    destRect.w = static_cast<int>(transform->width * transform->scale);
+    destRect.h = static_cast<int>(transform->height * transform->scale);
 }
